refactor(userContent): consolidate post + event stores into single useUserContentStore
The three-store layout (usePostStore, useEventStore, useUserContentStore)
collapses into one. The unified store now owns:

- myContent / fetchMyContent / upsert / remove (the unified list)
- createPost / updatePost / deletePost / setPostVisibility +
  hidePost/showPost (post CRUD, mirrors writes into myContent)
- createEvent / updateEvent / deleteEvent (event CRUD, same)
- fetchPublicPost (single-post detail, post-only public read)
- fetchPostsInBounds / postSummaries (post-only map summaries —
  parked here until the browse map UI gains event-aware markers
  and a clean rename to fetchInBounds takes over)

Method naming stays kind-prefixed for write/detail actions
(`createPost`, `fetchPublicPost`, …) so call sites read clearly;
unsuffixed for unified-list helpers (`fetchMyContent`, `upsert`).

Removed:
- usePostStore + spec; useEventStore (no spec)
- Cross-store `useUserContentStore().upsert(post)` calls inside the
  per-kind CRUD actions (now plain `this.upsert(post)`)
- Dead `fetchOwnerPost` and `currentPost` state — no consumers
- The empty `apps/frontend/src/features/{posts,events}/stores/`
  directories

Migrated consumers:
- EditPostDialog, EditEventDialog, PostsOrchestrator,
  BrowseProfiles, useBrowseViewModel — all import only
  useUserContentStore now
- BrowseProfiles + useBrowseViewModel test mocks updated

Net diff: +204 / -474 lines.

userContentStore.spec rewritten with 14 cases covering
fetchMyContent pagination, getters, upsert/remove, post CRUD,
event CRUD, and bounds — replacing the old postStore.spec.
